Added files for gettig acces only for an admin
