+++ packages/server/api/controllers/contactUsForm.controller.js
@@ -1,15 +1,15 @@
 const knex = require('../../config/db');
 
-const getContactFormData = async () => {
-  const contactFormData = knex('user_enquiry').select(
-    'user_enquiry.id',
-    'user_enquiry.first_name',
-    'user_enquiry.surname',
-    'user_enquiry.email',
-    'user_enquiry.message',
-  );
-  return contactFormData;
-};
+// const getContactFormData = async () => {
+//   const contactFormData = knex('user_enquiry').select(
+//     'user_enquiry.id',
+//     'user_enquiry.first_name',
+//     'user_enquiry.surname',
+//     'user_enquiry.email',
+//     'user_enquiry.message',
+//   );
+//   return contactFormData;
+// };
 
 const addDataFromContactForm = async (body) => {
   await knex('user_enquiry').insert({
@@ -25,6 +25,6 @@ const addDataFromContactForm = async (body) => {
 };
 
 module.exports = {
-  getContactFormData,
+  // getContactFormData,
   addDataFromContactForm,
 };
