+++ packages/server/firebase/serviceAccountKey.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "class21-fp-toolbox",
+  "private_key_id": "ef9a9c5a40fae727289f2e9acbcd5efcb7755eb6",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCkJuXdREc7exyW\nLS6tkr+vCY8mXFAafXj9daJQaoqkn9HLQFkOw+bgZhjH+EN1JJ10kpkt2enhiSln\nSaJPDDlMy7B/UDqx6ZTO0TgJsvpb5hRwiJnCZh3ErjbX82o1+/5jd1sv5/PoGgeI\nFK46cdiQQXX2YXIq0tMrsAl9WZB3dEzJO8GX3q1eTZIfobf28gwyCWc3uXeR0qrQ\nRstmIhDDs++inoOhe/lrSyVkm0RwAByUJQOBZTzYSXDAJ8ufXvaA8rZg5nk9aAWZ\nNU7D3fiGChSnquyCIK9914p2YcT1RjdAE0rNbWMhaDUVXNeTIsd1FAPfz8QIEuQt\nYPVM57k9AgMBAAECggEACawia/Ojqt/p4vjN66ZeVccKFf2R2W2+nQNlyWKYbB41\nXh+a38wv+ir/QiaL0Fw+9ukoeNOCp3/N3msv/YO/4JmOpdcpY5Ba7LDWoiOsrZDq\nGr4XKVBGaBodMIOu+iKww6r7HDXeYQp079IgI2DP8k1ue6e+8sJcymc4t+0jc8EL\nW8P9PfemPPqMXGwmxHF0tenZEHEvAhENa6wOoIdJBNgruuhX7edIfDC02Exb4BsU\nlAtBRQu1oRCjgMVVlZqoUcglrMAKnaYwQAQ+lXFG9JsKeucwux6nQ83Vfy7DIH73\nw88fbEOZt3+epZuHgniXMlzW8h1sPNVjomrYmpAzAQKBgQDdXixoZho2u2rea4sW\nAoUULeTHAFbYuEjNHM2BI4SNOvuoDRday5novTTFwXdzOsoLPyODY4RWjsmMhLNu\ng/CVgcsqauVsDRg9NrCC586ZmmhDYS1/UbNcxHCYCrVm0fuvEKU5GmOn0ikA0CpC\nkcsQX2qQ0BvdWoSHKrGzTxdBwQKBgQC91TcwlP4aLsbycLklv9GXP6iuph8mIuuj\nLtW0SiJfQ251Ndc6zndf4/LQ+akNymUGGSeZESbWdUd2ATGxqAs7d9H/vJ0O2C1z\nKcL2Ct8xoNUIrXbX900wnpddZMindmTJ5WN4Et+dcwAgzzUG8KyodYptGReD+S0x\nBvCN+tgefQKBgQCsDzWfE7VOU1AWfyqFOgIsFBvdHYNjQaG8Y6XLuUqe53iRP1Om\ny9feEapCB0UL+GbwZOv6hNBo4uT/Zml8/epymobKennGCUuM0MRTi0+HDn/51z+g\n1WT/NuR5GOG0gIQKp3p0rvFEkNI0ApjGU0GtgSx4UHlIDMIzuViTkWb3QQKBgEEK\n6CEOO85Ozr3M37NOeQrAthQ5OR0BBAnihziB2isX2KXyb69nhMFnQhBWKpVeDqU7\nQzUQCmWaw4RsPlGnyUlz8ZEVWB749QGrgnB5IyZG/BdAECNJSbpp7GUEnje6uifv\nH9jJZrL070tOk9Wv6wjBVU+yNDt0ygYtYYjUNO6lAoGBAMeogoBwpx1XHraG8U09\n+ihfxLKl88iT1fOFuUBCsQn4yeRdCjrg37g6+f6egmGYFgUGQ2B/yAOVQcjkTb6a\nwuFlQ61q6ZTfmcVLRWl7EIrHn6RiZOJHmD4OTAig+Vooc2mEds3zx/fhWEeriwni\n9M6D+wSnJXfXiofP/kG/tzkq\n-----END PRIVATE KEY-----\n",
+  "client_email": "firebase-adminsdk-nlthu@class21-fp-toolbox.iam.gserviceaccount.com",
+  "client_id": "109622051128019863909",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-nlthu%40class21-fp-toolbox.iam.gserviceaccount.com"
+}
